+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to the "veknamerge" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.4] - 2020-10-23
 
-- Initial release
+### Added
+
+* select your "ON" columns (key columns) used in the merge statement
+* added button/switch for APPLY TO ALL merge statements on those key columns
+* added more console out (for debugging)
+
+## [0.0.2] - 2020-10-22
+
+Mutliply insert statements and merge conversion
+
+## [0.0.1] - 2020-10-21
+
+### Added 
+
+Initial release. Single insert statement and merge conversion
+
+
+